Move permissions to the root dir
