+++ packages/snaps-rpc-methods/src/index.ts
@@ -7,4 +7,4 @@ export * from './restricted';
 export { SnapCaveatType } from '@metamask/snaps-utils';
 export { selectHooks } from './utils';
 export * from './endowments';
-export * from './permission-utils';
+export * from './permissions';
