--- packages/snaps-rpc-methods/src/permission-utils/index.ts
@@ -1,1 +1,0 @@
-export * from './permissions';
