+++ packages/snaps-rpc-methods/src/permissions.test.ts
